fix(mice_behavior): fix to_device() partial-init bug, halve PCIe transfer via late fp16->fp32 cast
Two fixes found while running the real final patch-grid training job:
1. to_device() set self.device on its first line, before allocating any GPU
   tensor. If a LATER allocation OOM'd (as it did for train_data on a 2080ti
   with only 11GB VRAM vs the ~16GB train array), the object was left
   half-initialized: self.device set (non-None) but self.offsets_grid_t (and
   others after the failing line) never created — get_batch then crashed
   with AttributeError instead of falling back to CPU gather. Fixed by using
   local variables throughout and only assigning to self once everything
   succeeds, so a mid-way OOM leaves the object in its original clean state.
2. get_batch's CPU/numpy path was upcasting fp16->fp32 before returning,
   which the caller then transferred to GPU — doubling both the CPU gather
   volume and the host->device PCIe payload (~49GB->~98GB/epoch for
   patch-grid) for no benefit. Now returns native dtype; the .float() upcast
   happens in train_fast()'s loop AFTER .to(dev), where it's GPU-memory-
   bandwidth-bound instead of PCIe-bound (and a harmless no-op for the
   already-fp32 GPU-resident path). Correctness re-verified: exact match
   (max abs diff 0.0) against the old dataset on 100 random samples.

+++ src/mice_behavior/train.py
@@ -353,8 +353,11 @@ def train_fast(
         for b0 in range(0, len(epoch_idx), batch_size):
             batch_idx = epoch_idx[b0:b0 + batch_size]
             ctx, offs, a1, a2, lbl, mask = train_data.get_batch(batch_idx)
+            # .float() after the device transfer, not before — upcasting fp16->fp32 pre-transfer
+            # would double the host->device PCIe payload for no benefit (GPU-resident path already
+            # returns fp32; this is then a harmless no-op there).
             ctx, offs, a1, a2, lbl, mask = (
-                ctx.to(dev, non_blocking=True), offs.to(dev, non_blocking=True),
+                ctx.to(dev, non_blocking=True).float(), offs.to(dev, non_blocking=True),
                 a1.to(dev, non_blocking=True), a2.to(dev, non_blocking=True),
                 lbl.to(dev, non_blocking=True), mask.to(dev, non_blocking=True),
             )
@@ -386,7 +389,7 @@ def train_fast(
                     batch_idx = val_keep[b0:b0 + batch_size]
                     ctx, offs, a1, a2, lbl, mask = val_data.get_batch(batch_idx)
                     ctx, offs, a1, a2, lbl, mask = (
-                        ctx.to(dev, non_blocking=True), offs.to(dev, non_blocking=True),
+                        ctx.to(dev, non_blocking=True).float(), offs.to(dev, non_blocking=True),
                         a1.to(dev, non_blocking=True), a2.to(dev, non_blocking=True),
                         lbl.to(dev, non_blocking=True), mask.to(dev, non_blocking=True),
                     )
